+++ packages/types/src/appearance.ts
@@ -170,6 +170,7 @@ export type ElementsConfig = {
   header: WithOptions;
   headerTitle: WithOptions;
   headerSubtitle: WithOptions;
+  headerBackLink: WithOptions;
 
   backRow: WithOptions;
   backLink: WithOptions;
@@ -462,6 +463,25 @@ export type ElementsConfig = {
   paymentSourceRowType: WithOptions;
   paymentSourceRowValue: WithOptions;
   paymentSourceRowBadge: WithOptions<'default' | 'expired'>;
+
+  invoiceRoot: WithOptions;
+  invoiceCard: WithOptions;
+  invoiceHeader: WithOptions;
+  invoiceHeaderContent: WithOptions;
+  invoiceTitle: WithOptions;
+  invoiceHeaderTitleBadgeContainer: WithOptions;
+  invoiceTitleIdContainer: WithOptions;
+  invoiceId: WithOptions;
+  invoiceIdContainer: WithOptions;
+  invoiceBadge: WithOptions;
+  invoiceDetails: WithOptions;
+  invoiceDetailsItem: WithOptions;
+  invoiceDetailsItemTitle: WithOptions;
+  invoiceDetailsItemTitleText: WithOptions;
+  invoiceDetailsItemValue: WithOptions;
+  invoiceDetailsItemValueText: WithOptions;
+  invoiceCopyButton: WithOptions;
+  invoiceContent: WithOptions;
 
   menuButton: WithOptions<MenuId>;
   menuButtonEllipsis: WithOptions;