+++ packages/types/src/signIn.ts
@@ -72,7 +72,6 @@ import type { AuthenticateWithWeb3Params } from './web3Wallet';
 
 /**
  * The `SignIn` object holds the state of the current sign-in and provides helper methods to navigate and complete the sign-in process. It is used to manage the sign-in lifecycle, including the first and second factor verification, and the creation of a new session.
- * @interface
  */
 export interface SignInResource extends ClerkResource {
   /**
@@ -120,6 +119,9 @@ export interface SignInResource extends ClerkResource {
   createEmailLinkFlow: () => CreateEmailLinkFlowReturn<SignInStartEmailLinkFlowParams, SignInResource>;
 
   validatePassword: (password: string, callbacks?: ValidatePasswordCallbacks) => void;
+  /**
+   * @internal
+   */
   __internal_toSnapshot: () => SignInJSONSnapshot;
 }