+++ packages/chrome-extension/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 2.2.30
+
+### Patch Changes
+
+- Updated dependencies [[`5ce5147`](https://github.com/clerk/javascript/commit/5ce5147adec8835b15f974e7210b775e4667a3aa), [`de18497`](https://github.com/clerk/javascript/commit/de184970818dca898b67c0fa1277a69b123b5e93), [`7389ba3`](https://github.com/clerk/javascript/commit/7389ba3164ca0d848fb0a9de5d7e9716925fadcc), [`59284dc`](https://github.com/clerk/javascript/commit/59284dc9f03d40633c8b1b4079fbf3d051776c7b), [`27ca2ec`](https://github.com/clerk/javascript/commit/27ca2ece648b346d044f9a65890e34303cee9042), [`f6ef841`](https://github.com/clerk/javascript/commit/f6ef841125ff21ca8cae731d1f47f3a101d887e1), [`e634830`](https://github.com/clerk/javascript/commit/e6348301ab56a7868f24c1b9a4dd9e1d60f6027b), [`f8887b2`](https://github.com/clerk/javascript/commit/f8887b2cbd145e8e49bec890e8b6e02e34178d6a)]:
+  - @clerk/clerk-js@5.61.1
+  - @clerk/shared@3.7.1
+  - @clerk/clerk-react@5.28.1
+
 ## 2.2.29
 
 ### Patch Changes